+++ psakeFile.ps1
@@ -1,7 +1,7 @@
 # Initialize the BuildHelpers environment variables here so they are usable in all child scopes including the psake properties block
 BuildHelpers\Set-BuildEnvironment -Force
 
-properties {
+Properties {
 
     # Version of the module manifest in the src directory before the build is run and the version is updated
     $SourceModuleVersion = (Import-PowerShellDataFile -Path $env:BHPSModuleManifest).ModuleVersion
@@ -92,18 +92,18 @@ FormatTaskName {
     Write-Host $taskName -ForegroundColor Blue
 }
 
-task Default -depends Publish
+Task Default -depends Publish
 
 #Task Init -FromModule PowerShellBuild -minimumVersion 0.6.1
 
-task InitializeEnvironmentVariables {
+Task InitializeEnvironmentVariables {
 
     # Should I be running Git before this? I haven't run Git yet, so BuildHelpers finds the previous commit msg and I have to use the line below to update it
     $env:BHCommitMessage = $CommitMessage
 
 } -description 'Initialize the environment variables from the BuildHelpers module'
 
-task UpdateModuleVersion -depends InitializeEnvironmentVariables -Action {
+Task UpdateModuleVersion -depends InitializeEnvironmentVariables -action {
     $CurrentVersion = (Test-ModuleManifest $env:BHPSModuleManifest).Version
     "`tOld Version: $CurrentVersion"
     if ($IncrementMajorVersion) {
@@ -121,7 +121,7 @@ task UpdateModuleVersion -depends InitializeEnvironmentVariables -Action {
     Update-Metadata -Path $env:BHPSModuleManifest -PropertyName ModuleVersion -Value $NewModuleVersion -ErrorAction Stop
 } -description 'Increment the module version and update the module manifest accordingly'
 
-task InitializePowershellBuild -depends UpdateModuleVersion {
+Task InitializePowershellBuild -depends UpdateModuleVersion {
 
     $NewModuleVersion = (Import-PowerShellDataFile -Path $env:BHPSModuleManifest).ModuleVersion
 
@@ -161,7 +161,7 @@ task InitializePowershellBuild -depends UpdateModuleVersion {
 
 } -description 'Initialize environment variables from the PowerShellBuild module'
 
-task RotateBuilds -depends InitializePowershellBuild {
+Task RotateBuilds -depends InitializePowershellBuild {
     $BuildVersionsToRetain = 1
     Get-ChildItem -Directory -Path $BuildOutDir |
     Sort-Object -Property Name |
@@ -173,7 +173,7 @@ task RotateBuilds -depends InitializePowershellBuild {
     $NewLine
 } -description 'Delete all but the last 4 builds, so we will have our 5 most recent builds after the new one is complete'
 
-task UpdateChangeLog -depends RotateBuilds -Action {
+Task UpdateChangeLog -depends RotateBuilds -action {
     <#
 TODO
     This task runs before the Test task so that tests of the change log will pass
@@ -184,7 +184,7 @@ TODO
 #>
     $ChangeLog = "$env:BHProjectPath\CHANGELOG.md"
     $NewModuleVersion = (Import-PowerShellDataFile -Path $env:BHPSModuleManifest).ModuleVersion
-    $NewChanges = "## [$NewModuleVersion] - $(Get-Date -format 'yyyy-MM-dd') - $CommitMessage$NewLine"
+    $NewChanges = "## [$NewModuleVersion] - $(Get-Date -Format 'yyyy-MM-dd') - $CommitMessage$NewLine"
     "`tChange Log:  $ChangeLog"
     "`tNew Changes: $NewChanges"
     [string[]]$ChangeLogContents = Get-Content -Path $ChangeLog
@@ -200,7 +200,7 @@ TODO
     $NewChangeLogContents | Out-File -FilePath $ChangeLog -Encoding utf8 -Force
 }
 
-task ExportPublicFunctions -depends UpdateChangeLog -Action {
+Task ExportPublicFunctions -depends UpdateChangeLog -action {
     # Discover public functions
     $ScriptFiles = Get-ChildItem -Path "$env:BHPSModulePath\*.ps1" -Recurse
     $PublicScriptFiles = $ScriptFiles | Where-Object -FilterScript {
@@ -222,17 +222,17 @@ task ExportPublicFunctions -depends UpdateChangeLog -Action {
     }
 
     # Export public functions in the manifest
-    Update-MetaData -Path $env:BHPSModuleManifest -PropertyName FunctionsToExport -Value $publicFunctions
+    Update-Metadata -Path $env:BHPSModuleManifest -PropertyName FunctionsToExport -Value $publicFunctions
 
 } -description 'Export all public functions in the module'
 
-task CleanOutputDir -depends ExportPublicFunctions {
+Task CleanOutputDir -depends ExportPublicFunctions {
     "`tOutput: $env:BHBuildOutput"
     Clear-PSBuildOutputFolder -Path $env:BHBuildOutput
     $NewLine
 } -description 'Clears module output directory'
 
-task BuildModule -depends CleanOutputDir {
+Task BuildModule -depends CleanOutputDir {
     $buildParams = @{
         Path               = $env:BHPSModulePath
         ModuleName         = $env:BHProjectName
@@ -271,14 +271,14 @@ $genMarkdownPreReqs = {
     $result
 }
 
-task DeleteMarkdownHelp -depends BuildModule -precondition $genMarkdownPreReqs {
+Task DeleteMarkdownHelp -depends BuildModule -precondition $genMarkdownPreReqs {
     $MarkdownDir = [IO.Path]::Combine($DocsRootDir, $HelpDefaultLocale)
     "`tDeleting folder: '$MarkdownDir'"
     Get-ChildItem -Path $MarkdownDir -Recurse | Remove-Item
     $NewLine
 } -description 'Delete existing .md files to prepare for PlatyPS to build new ones'
 
-task BuildMarkdownHelp -depends DeleteMarkdownHelp {
+Task BuildMarkdownHelp -depends DeleteMarkdownHelp {
     $ManifestPath = [IO.Path]::Combine($env:BHBuildOutput, "$env:BHProjectName.psd1")
     $moduleInfo = Import-Module $ManifestPath  -Global -Force -PassThru
     $manifestInfo = Test-ModuleManifest -Path $ManifestPath
@@ -316,7 +316,7 @@ task BuildMarkdownHelp -depends DeleteMarkdownHelp {
     }
 } -description 'Generate markdown files from the module help'
 
-task FixMarkdownHelp -depends BuildMarkdownHelp {
+Task FixMarkdownHelp -depends BuildMarkdownHelp {
     $ManifestPath = [IO.Path]::Combine($env:BHBuildOutput, "$env:BHProjectName.psd1")
     $moduleInfo = Import-Module $ManifestPath  -Global -Force -PassThru
     $manifestInfo = Test-ModuleManifest -Path $ManifestPath
@@ -326,7 +326,7 @@ task FixMarkdownHelp -depends BuildMarkdownHelp {
     [string]$ModuleHelp = Get-Content -LiteralPath $ModuleHelpFile -Raw
 
     #-Update the module description
-    $RegEx = "(?ms)\#\#\ Description\s*[^\r\n]*\s*"
+    $RegEx = '(?ms)\#\#\ Description\s*[^\r\n]*\s*'
     $NewString = "## Description$NewLine$($moduleInfo.Description)$NewLine$NewLine"
     $ModuleHelp = $ModuleHelp -replace $RegEx, $NewString
 
@@ -334,7 +334,7 @@ task FixMarkdownHelp -depends BuildMarkdownHelp {
 
     #-Update the description of each function (use its synopsis for brevity)
     ForEach ($ThisFunction in $ManifestInfo.ExportedCommands.Keys) {
-        $Synopsis = (Get-Help -Name $ThisFunction).Synopsis
+        $Synopsis = (Get-Help -name $ThisFunction).Synopsis
         $RegEx = "(?ms)\#\#\#\ \[$ThisFunction]\($ThisFunction\.md\)\s*[^\r\n]*\s*"
         $NewString = "### [$ThisFunction]($ThisFunction.md)$NewLine$Synopsis$NewLine$NewLine"
         $ModuleHelp = $ModuleHelp -replace $RegEx, $NewString
@@ -353,7 +353,7 @@ $genHelpFilesPreReqs = {
     $result
 }
 
-task BuildMAMLHelp -depends FixMarkdownHelp -precondition $genHelpFilesPreReqs {
+Task BuildMAMLHelp -depends FixMarkdownHelp -precondition $genHelpFilesPreReqs {
     Build-PSBuildMAMLHelp -Path $DocsRootDir -DestinationPath $env:BHBuildOutput
 } -description 'Generates MAML-based help from PlatyPS markdown files'
 
@@ -366,7 +366,7 @@ $genUpdatableHelpPreReqs = {
     $result
 }
 
-task BuildUpdatableHelp -depends BuildMAMLHelp -precondition $genUpdatableHelpPreReqs {
+Task BuildUpdatableHelp -depends BuildMAMLHelp -precondition $genUpdatableHelpPreReqs {
 
     $OS = (Get-CimInstance -ClassName CIM_OperatingSystem).Caption
     if ($OS -notmatch 'Windows') {
@@ -415,7 +415,7 @@ $analyzePreReqs = {
     $result
 }
 
-task Lint -depends BuildUpdatableHelp -precondition $analyzePreReqs {
+Task Lint -depends BuildUpdatableHelp -precondition $analyzePreReqs {
     $analyzeParams = @{
         Path              = $env:BHBuildOutput
         SeverityThreshold = $TestLintFailBuildOnSeverityLevel
@@ -441,7 +441,10 @@ $pesterPreReqs = {
     return $result
 }
 
-task UnitTests -depends Lint -precondition $pesterPreReqs {
+Task UnitTests -depends Lint -precondition $pesterPreReqs {
+
+    Write-Host "`tInvoke-Pester -Configuration `$PesterConfiguration"
+
     $pesterParams = @{
         Path                         = $TestRootDir
         ModuleName                   = $env:BHProjectName
@@ -455,17 +458,18 @@ task UnitTests -depends Lint -precondition $pesterPreReqs {
         CodeCoverageOutputFileFormat = $TestCodeCoverageOutputFormat
         ImportModule                 = $TestImportModuleFirst
     }
-    Test-PSBuildPester @pesterParams
+    #Test-PSBuildPester @pesterParams
+
 } -description 'Execute Pester tests'
 
-task SourceControl -depends UnitTests {
+Task SourceControl -depends UnitTests {
     # Commit to Git
     git add .
     git commit -m $CommitMessage
     git push origin main
 } -description 'git add, commit, and push'
 
-task Publish -depends SourceControl {
+Task Publish -depends SourceControl {
     Assert -conditionToCheck ($PublishPSRepositoryApiKey -or $PublishPSRepositoryCredential) -failureMessage "API key or credential not defined to authenticate with [$PublishPSRepository)] with."
 
     $publishParams = @{
@@ -485,14 +489,14 @@ task Publish -depends SourceControl {
     Publish-Module @publishParams
 } -description 'Publish module to the defined PowerShell repository'
 
-task FinalTasks -depends Publish {
+Task FinalTasks -depends Publish {
 
     # Remove script-scoped variables to avoid their accidental re-use
-    Remove-Variable -Name ModuleOutDir -Scope Script -Force -ErrorAction SilentlyContinue
+    Remove-Variable -name ModuleOutDir -Scope Script -Force -ErrorAction SilentlyContinue
 
 }
 
-task ? -description 'Lists the available tasks' {
+Task ? -description 'Lists the available tasks' {
     'Available tasks:'
     $psake.context.Peek().Tasks.Keys | Sort-Object
 }